Add additional information in the url_rule docs (#2467)
If a user is making use of the errorhandler(405) decorator, and
they wish to include an Allow header (as per HTTP spec), they may
be confused to find that url_rule is None. This doc change aims to
clarify that because the request was never successfully matched,
it was never bound to a URL rule and to access the valid methods,
they must examine routing_exception.