Make the documentation build reproducibly
Whilst working on the Reproducible Builds effort [0] we noticed that
flask could not be built reproducibly.

This is because it includes an absolute build directory in the
documentation as the "json_module" attribute points to a Python class/
module which has a string representation including its path.

This commit skips this (inherited) member from the documentation.

(This was originally filed in Debian as #943674 [1].)

 [0] https://reproducible-builds.org/
 [1] https://bugs.debian.org/943674